+++ TowerDefense/Assets/Scripts/Algorithms/GeneticAlgorithm/Heuristics/HeuristicsCalculator.cs
@@ -5,7 +5,7 @@ namespace TowerDefense.Algorithms.GeneticAlgorithm.Heuristics
 {
     public abstract class HeuristicsCalculator<TGene>
     {
-        public abstract int GetHeuristicsValue(List<TGene> genes);
+        public abstract float GetHeuristicsValue(List<TGene> genes);
         
     }
 }
